various changes to clean warnings from MSVC builds
git-svn-id: https://svn.osgeo.org/gdal/trunk/gdal@21680 f0d54148-0727-0410-94bb-9a71ac55c965

+++ frmts/hdf5/gh5_convenience.cpp
@@ -118,7 +118,7 @@ bool GH5_FetchAttribute( hid_t loc_id, const char *pszAttrName,
     int i, nAttrElements = 1;
 
     for( i=0; i < nAttrDims; i++ ) {
-        nAttrElements *= anSize[i];
+        nAttrElements *= (int) anSize[i];
     }
 
     if( nAttrElements != 1 )
